pyproject.toml
[project]
name = "project5-aiml-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4",
    "gradio",
    "langchain",
    "langchain-classic",
    "langchain-community",
    "langchain-core",
    "langchain-groq",
    "langchain-huggingface",
    "langchain-ollama",
    "langchain-pinecone",
    "lxml",
    "pinecone",
    "pydantic",
    "pypdf",
    "python-dotenv",
    "sentence-transformers",
    "sse-starlette",
]
